Rename from mllib to ml for the new added classes, and also fixed the SparkFunSuite imports
