+++ mllib-local/src/main/scala/org/apache/spark/ml/linalg/BLAS.scala
@@ -15,17 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.spark.mllib.linalg
+package org.apache.spark.ml.linalg
 
 import com.github.fommil.netlib.{BLAS => NetlibBLAS, F2jBLAS}
 import com.github.fommil.netlib.BLAS.{getInstance => NativeBLAS}
 
-import org.apache.spark.internal.Logging
-
 /**
  * BLAS routines for MLlib's vectors and matrices.
  */
-private[spark] object BLAS extends Serializable with Logging {
+private[spark] object BLAS extends Serializable {
 
   @transient private var _f2jBLAS: NetlibBLAS = _
   @transient private var _nativeBLAS: NetlibBLAS = _
