+++ mllib-local/src/main/scala/org/apache/spark/ml/linalg/Matrices.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.mllib.linalg
+package org.apache.spark.ml.linalg
 
 import java.util.{Arrays, Random}
