+++ mllib-local/src/test/scala/org/apache/spark/ml/linalg/BreezeMatrixConversionSuite.scala
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.spark.mllib.linalg
+package org.apache.spark.ml.linalg
 
 import breeze.linalg.{CSCMatrix => BSM, DenseMatrix => BDM}
 
-import org.apache.spark.SparkFunSuite
+import org.apache.spark.ml.SparkMLFunSuite
 
-class BreezeMatrixConversionSuite extends SparkFunSuite {
+class BreezeMatrixConversionSuite extends SparkMLFunSuite {
   test("dense matrix to breeze") {
     val mat = Matrices.dense(3, 2, Array(0.0, 1.0, 2.0, 3.0, 4.0, 5.0))
     val breeze = mat.toBreeze.asInstanceOf[BDM[Double]]
